refactor: extract setup fields section
